Add check if there are results presents. Throw argumentException

+++ PimPamProgrammeur.Backend/PimPamProgrammeur.API/Processors/ClassroomProcessor.cs
@@ -15,12 +15,14 @@ namespace PimPamProgrammeur.API.Processors
         private readonly IMapper _mapper;
         private readonly IClassroomRepository _classroomRepository;
         private readonly IUserRepository _userRepository;
+        private readonly IResultRepository _resultRepository;
 
-        public ClassroomProcessor(IClassroomRepository classroomRepository, IMapper mapper, IUserRepository userRepository)
+        public ClassroomProcessor(IClassroomRepository classroomRepository, IMapper mapper, IUserRepository userRepository, IResultRepository resultRepository)
         {
             _classroomRepository = classroomRepository;
             _mapper = mapper;
             _userRepository = userRepository;
+            _resultRepository = resultRepository;
         }
 
         public async Task DeleteClassroom(Guid id)
@@ -28,8 +30,14 @@ namespace PimPamProgrammeur.API.Processors
             if (_classroomRepository.GetClassroom(id) != null)
             {
                 var getAllUsers = _userRepository.GetUserByClassroomId(id);
+
                 if (getAllUsers.Any())
                 {
+                    var usersCantBeDeleted = CheckIfUsersCanBeDeleted(getAllUsers);
+                    if (usersCantBeDeleted)
+                    {
+                        throw new ArgumentException("There are users with results, first delete the results");
+                    }
                     await _userRepository.DeleteAllUsers(getAllUsers);
                 }
                 await _classroomRepository.DeleteClassroom(id);
@@ -57,6 +65,21 @@ namespace PimPamProgrammeur.API.Processors
             var classroomResponse =  _mapper.Map<ClassroomResponseDto>(savedClassroom);
 
             return classroomResponse;
+        }
+
+        public bool CheckIfUsersCanBeDeleted(IEnumerable<User> users)
+        {
+            var usersHasResults = false;
+            foreach (var user in users)
+            {
+               var hasResults = _resultRepository.CheckIfResultsArePresentByUserId(user.Id);
+                if (hasResults)
+                {
+                    usersHasResults = true;
+                    return usersHasResults;
+                }
+            }
+            return usersHasResults;
         }
     }
 }
